requirements.txt
Django==5.1.2
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
django-storages==1.14.4
boto3==1.35.44
Pillow==10.1.0
PyPDF2==3.0.1
reportlab==4.0.9
docx2pdf==0.1.8
python-docx==1.1.2
